move install unit tests related to bin linking to own file
